move managed client code to com.teragrep.rlp_01.client
