+++ src/main/java/com/teragrep/rlp_01/client/ManagedRelpConnectionStub.java
@@ -14,9 +14,27 @@
    See the License for the specific language governing permissions and
    limitations under the License.
  */
-package com.teragrep.rlp_01.pool;
+package com.teragrep.rlp_01.client;
 
-public interface IManagedRelpConnection extends Poolable {
-    void reconnect();
-    void ensureSent(byte[] bytes);
+public class ManagedRelpConnectionStub implements IManagedRelpConnection {
+
+    @Override
+    public void reconnect() {
+        throw new IllegalStateException("ManagedRelpConnectionStub does not support this");
+    }
+
+    @Override
+    public void ensureSent(byte[] bytes) {
+        throw new IllegalStateException("ManagedRelpConnectionStub does not support this");
+    }
+
+    @Override
+    public boolean isStub() {
+        return true;
+    }
+
+    @Override
+    public void close() {
+        throw new IllegalStateException("ManagedRelpConnectionStub does not support this");
+    }
 }
