+++ src/main/java/com/teragrep/rlp_01/client/RebindableRelpConnection.java
@@ -14,7 +14,7 @@
    See the License for the specific language governing permissions and
    limitations under the License.
  */
-package com.teragrep.rlp_01.pool;
+package com.teragrep.rlp_01.client;
 
 import java.io.IOException;
