+++ src/main/java/com/teragrep/rlp_01/client/RelpConnectionWithConfig.java
@@ -14,7 +14,7 @@
    See the License for the specific language governing permissions and
    limitations under the License.
  */
-package com.teragrep.rlp_01.pool;
+package com.teragrep.rlp_01.client;
 
 import com.teragrep.rlp_01.RelpBatch;
 import com.teragrep.rlp_01.RelpConnection;
