+++ src/main/java/com/teragrep/rlp_01/client/RenewableRelpConnection.java
@@ -1,4 +1,4 @@
-package com.teragrep.rlp_01.pool;
+package com.teragrep.rlp_01.client;
 
 import java.io.IOException;
 import java.time.Instant;
